fixed addition of invalid buddies (there is a serious uint/int misuse here)

+++ trunk/src/server/bulkobjects/pscharacterloader.cpp
@@ -805,7 +805,7 @@ PID psCharacterLoader::FindCharacterID(const char *character_name, bool excludeN
         result = db->SelectSingleNumber("SELECT id from characters where name='%s' AND npc_master_id=0",escape.GetData());
     }
 
-    return PID(result == QUERY_FAILED ? 0 : result);
+    return PID((int)result == QUERY_FAILED ? 0 : result);
 }
 
 PID psCharacterLoader::FindCharacterID(AccountID accountID, const char *character_name)
